requirements.txt
auto_mix_prep==0.2.0
fastapi==0.115.0
pydantic==2.9.2
python-dotenv==1.0.1
uvicorn==0.31.0
